Darken normal text color, remove line coloring
I've darkened the default normal text color and removed cursorline text
darkening. On the whole, the light grey is too difficult to read against
a white background, and the line by line focus provided by the
cursorline darkening proved annoying and unhelpful to me, especially
when I had to read source code using this scheme (if I weren't lazy, I'd
switch to an alternative color scheme in such situations, but I don't
want to have to do that for quick browsing).